+++ DragonbornPresence.h
@@ -4,6 +4,7 @@
 
 namespace DragonbornPresence {
 
+void LoadConfig();
 void SetLocale();
 void RegisterGameEventHandlers();
 void OnGameLoaded();